Switching to AttributeRef and precomputing values
All in the name of speed. The HDK docs indicate there might be some
overhead in the nice clean AttributeHandle API so we use AttributeRefs
instead.

Then we precompute all the necessary vertex positions and indicies for
drawing quads quickly without fussing around with glTransforms and
starting and stopping GL_QUADS.

Seems to have improved things a little.