Changed discs to face the camera
We achieve this by querying the modelview matrix and taking the
translate components and treating that as the camera position, then
calculating vectors to it, and orthogonal to it from the particle
positions and using them as our basis for drawing the discs.

Currently ignores scale and rotate. The former should be fixed before
long.