Taken up vector from camera and changed use of scale
Instead of trying to hack our own up vector which was flawed as you
never know which way your facing, we take one from the camera matrix.

We scale differently depending on whether we're camera facing or not.
For camera facing we have to apply the scale, x to the "left" and y to
the "up", after they have been calculated. For normal rotations, we
scale first so that we don't get confused by the rotation and skew it.