Refactored into multiple files
Slightly reducing the ugliness of it by creating separate Renderer and
DisplayObject files and factories. Not perfect but it is a start.

Extended the build files to match the new structure.

Forced to define NEED_SPECIALIZATION_STORAGE on the compile line to stop
linking errors with g++-4.1. I've no idea if this is required for more
up to date compilers.